update Model::find to handle cases when $id is 0
Only the case where $id was set to NULL was being tested. And in the
code the condition of the if statement was simply if ($id) { ... }.

However in PHP 0 is a falsy value (evaluate to false in an if statment)
so it wouldn't actually search for a record that had it 0.

This was
fixed using is_null in the condition instead.